Rename seqfiles cli module
